requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
pypdf>=4.0.0
groq>=0.13.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
tabulate>=0.9.0
rich>=13.7.0
pytest>=8.0.0
reportlab>=4.0.0
